add cach warming to `regions.const`
