+++ app/src/app/regionen/[regionSlug]/_mapData/mapDataSources/generalization/generalizationIdentifier.ts
@@ -1,23 +1,5 @@
-// This type includes all tables we generate in atlas-geo
-// TODO: automatically generate this in atlas-geo
-export type TableId =
-  | 'barrierLines'
-  | 'barrierAreas'
-  | 'bicycleParking_points'
-  | 'bicycleParking_areas'
-  | 'bikeroutes'
-  | 'boundaries'
-  | 'boundaryLabels'
-  | 'presenceStats'
-  | 'landuse'
-  | 'places'
-  | 'poiClassification'
-  | 'publicTransport'
-  | 'roads'
-  | 'roadsPathClasses'
-  | 'bikelanesPresence' // based on `roads`
-  | 'bikelanes'
-  | 'trafficSigns'
+import { TableId } from 'src/app/regionen/[regionSlug]/_mapData/mapDataSources/tables.const'
 
+export type GeneralizedTableId = `atlas_generalized_${Lowercase<TableId>}`
 export const generalizationFunctionIdentifier = (tableName: TableId) =>
-  `atlas_generalized_${tableName.toLowerCase()}` as `atlas_generalized_${Lowercase<TableId>}`
+  `atlas_generalized_${tableName.toLowerCase()}` as GeneralizedTableId
