+++ app/src/app/regionen/[regionSlug]/_mapData/mapDataSources/sources.const.ts
@@ -6,10 +6,7 @@ import { SourceExportApiIdentifier } from './export/exportIdentifier'
 import { sourcesParking, SourcesParkingId } from './sourcesParking.const'
 import { SourceVerificationApiIdentifier } from './verification/verificationIdentifier'
 
-// TODO type MapDataConfigSourcesIds = typeof sources[number]['id']
-export type SourcesId =
-  | SourcesParkingId
-  | 'accidents_unfallatlas'
+export type AtlasSourceId =
   | 'atlas_barriers'
   | 'atlas_bicycleParking'
   | 'atlas_bikelanes'
@@ -24,9 +21,18 @@ export type SourcesId =
   | 'atlas_roadsPathClasses'
   | 'atlas_bikelanesPresence' // based on `roads`
   | 'atlas_trafficSigns'
+
+export type MapillarySourceId =
   | 'mapillary_coverage'
   | 'mapillary_mapfeatures'
   | 'mapillary_trafficSigns'
+
+// TODO type MapDataConfigSourcesIds = typeof sources[number]['id']
+export type SourcesId =
+  | SourcesParkingId
+  | AtlasSourceId
+  | MapillarySourceId
+  | 'accidents_unfallatlas'
 
 export const sources: MapDataSource<
   SourcesId,
