+++ app/src/app/regionen/[regionSlug]/_mapData/mapDataSources/tables.const.ts
@@ -1,0 +1,19 @@
+// This is a list of all the tables we generate in atlas-geo. We might auto generate this list in the future.
+export type TableId =
+  | 'barrierLines'
+  | 'barrierAreas'
+  | 'bicycleParking_points'
+  | 'bicycleParking_areas'
+  | 'bikeroutes'
+  | 'boundaries'
+  | 'boundaryLabels'
+  | 'presenceStats'
+  | 'landuse'
+  | 'places'
+  | 'poiClassification'
+  | 'publicTransport'
+  | 'roads'
+  | 'roadsPathClasses'
+  | 'bikelanesPresence' // based on `roads`
+  | 'bikelanes'
+  | 'trafficSigns'
