+++ app/src/instrumentation/registerGeneralizationFunctions.ts
@@ -1,8 +1,6 @@
-import {
-  generalizationFunctionIdentifier,
-  TableId,
-} from '@/src/app/regionen/[regionSlug]/_mapData/mapDataSources/generalization/generalizationIdentifier'
+import { generalizationFunctionIdentifier } from '@/src/app/regionen/[regionSlug]/_mapData/mapDataSources/generalization/generalizationIdentifier'
 import { InteracitvityConfiguartion } from '@/src/app/regionen/[regionSlug]/_mapData/mapDataSources/generalization/interacitvityConfiguartion'
+import { TableId } from '@/src/app/regionen/[regionSlug]/_mapData/mapDataSources/tables.const'
 import { geoDataClient } from '@/src/prisma-client'
 
 async function createTileSpecification(tableName: TableId) {
